houduan/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
distro==1.9.0
docutils==0.22.4
exceptiongroup==1.3.1
filetype==1.2.0
Flask==3.1.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kivymd==1.2.0
MarkupSafe==3.0.3
openai==2.24.0
pillow==11.3.0
plyer==2.1.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypiwin32==223
pywin32==311
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
zipp==3.23.0
